added stacks folder with homeStack exported to it
